added error handling

+++ Application/Activities/Commands/DeleteActivity.cs
@@ -1,3 +1,4 @@
+using Application.Core;
 using MediatR;
 using Persistence;
 
@@ -5,19 +6,21 @@ namespace Application.Activities.Commands;
 
 public class DeleteActivity
 {
-	public class Command : IRequest
-	{
-		public required string Id { get; set; }
-	};
+    public class Command : IRequest<Result<Unit>>
+    {
+        public required string Id { get; set; }
+    };
 
-	public class Handler(AppDbContext context) : IRequestHandler<Command>
-	{
-		public async Task Handle(Command request, CancellationToken cancellationToken)
-		{
-			var activity = await context.Activities.FindAsync([request.Id], cancellationToken)
-				?? throw new KeyNotFoundException("Activity not found");
-			context.Activities.Remove(activity);
-			await context.SaveChangesAsync(cancellationToken);
-		}
-	}
+    public class Handler(AppDbContext context) : IRequestHandler<Command, Result<Unit>>
+    {
+        public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
+        {
+            var activity = await context.Activities.FindAsync([request.Id], cancellationToken);
+            if (activity == null) return Result<Unit>.Failure("Activity not found", 404);
+            context.Activities.Remove(activity);
+            return await context.SaveChangesAsync(cancellationToken) > 0
+                ? Result<Unit>.Success(Unit.Value)
+                : Result<Unit>.Failure("Failed to create activity", 400);
+        }
+    }
 }
